+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.4.4'
-__released__ = '2019-09-23'
+__version__ = '1.4.5'
+__released__ = '2019-09-27'
 
 from .settings import Settings as Emborg